Refactor. Single day only. Comments.

+++ src/Controller/API/TimeEntryController.php
@@ -67,6 +67,49 @@ class TimeEntryController extends AbstractController
         return new JsonResponse(['success' => true]);
     }
 
+    #[Route('/add/comment', name: 'add_comment', methods: 'POST')]
+    public function comment(Request $request, Security $security, ProjectRepository $projectRepository, UserTimeEntryRepository $userTimeEntryRepository): Response
+    {
+        $fieldId = $request->request->get('field');
+        $fieldValue = $request->request->get('value');
+
+        $fieldIdParts = explode('|', $fieldId);
+        if (2 !== count($fieldIdParts)) {
+            return $this->createErrorResponse('Invalid field id encountered');
+        }
+
+        [$projectId, $dateString] = $fieldIdParts;
+        $project = $projectRepository->find($projectId);
+        if (!$project) {
+            return $this->createErrorResponse('Could not find project');
+        }
+
+        try {
+            $entryDate = new DateTimeImmutable($dateString);
+        } catch (\Exception) {
+            return $this->createErrorResponse('Could not parse date');
+        }
+
+        $user = $security->getUser();
+        if (!$user instanceof User) {
+            return $this->createErrorResponse('Weird unknown user type thing');
+        }
+
+        $timeEntry = $userTimeEntryRepository->findOneBy(['user' => $user, 'project' => $project, 'entryDateInt' => $entryDate->format('Ymd')]);
+        if ($timeEntry && (int)$fieldValue === 0 && !$fieldValue) {
+            $userTimeEntryRepository->remove($timeEntry, true);
+        } else {
+            if (!$timeEntry) {
+                $timeEntry = new UserTimeEntry($user, $project, $entryDate);
+                $timeEntry->setHours(0);
+            }
+            $timeEntry->setComment($fieldValue);
+            $userTimeEntryRepository->save($timeEntry, true);
+        }
+
+        return new JsonResponse(['success' => true]);
+    }
+
     #[Route('/add', name: 'add', methods: 'POST')]
     public function add(Request $request, Security $security, ProjectRepository $projectRepository, UserTimeEntryRepository $userTimeEntryRepository): Response
     {
@@ -96,7 +139,7 @@ class TimeEntryController extends AbstractController
         }
 
         $timeEntry = $userTimeEntryRepository->findOneBy(['user' => $user, 'project' => $project, 'entryDateInt' => $entryDate->format('Ymd')]);
-        if ($timeEntry && (int)$fieldValue === 0) {
+        if ($timeEntry && (int)$fieldValue === 0 && !$timeEntry->getComment()) {
             $userTimeEntryRepository->remove($timeEntry, true);
         } else {
             if (!$timeEntry) {
